requirements.txt
Django==1.9
Pillow==3.2.0
django-thumbnailfield=0.2.5